@@ -4,19 +4,32 @@ version: 1.0.0
 publish_to: none
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: '>=2.17.1 <4.0.0'
+  flutter: ">=3.0.1"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
+
+  go_router: ^7.1.1
+  http: ^1.0.0
+  path: ^1.8.3
+
   upgrader:
     path: ../
 
 dev_dependencies:
+  flutter_driver:
+    sdk: flutter
+
+  flutter_lints: ^2.0.3
+
   flutter_test:
     sdk: flutter
+
+  test:
 
 flutter:
   uses-material-design: true
