@@ -1,52 +1,56 @@
 name: upgrader
 description: Flutter package for prompting users to upgrade when there is a newer version of the app in the store.
-version: 4.6.1
+version: 10.3.0
 homepage: https://github.com/larryaasen/upgrader
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=1.12.0"
+  sdk: '>=3.1.0 <4.0.0'
+  flutter: ">=3.16.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   # From fluttercommunity.dev: Get current device information from within the Flutter application.
-  device_info_plus: ">=3.2.0 <4.2.0"
-
-  # For app_theme.dart
-  google_fonts: ^2.3.1
-
-  # From Dart Team: A composable, Future-based library for making HTTP requests.
-  http: ^0.13.0
-
-  # From dart.dev: Platform independent access to information about the current operating system.
-  os_detect: ^2.0.0
-
-  # From Flutter Community: Provides an API for querying information about an application package.
-  package_info_plus: ^1.4.2
-
-  # From Flutter Team: Wraps NSUserDefaults (on iOS) and SharedPreferences (on Android).
-  shared_preferences: ^2.0.3
-
-  # From Flutter Team: A Flutter plugin for launching a URL in the mobile platform.
-  url_launcher: ^6.1.0
-
-  # A dart library for comparing and incrementing version numbers with Semantic Versioning.
-  version: ">=2.0.0 <3.1.0"
-
-  # A dart library for parsing, traversing, querying, transforming and building XML documents.
-  xml: ">=5.0.2 <7.0.0"
+  device_info_plus: '>=8.1.0 <11.0.0'
 
   # Pure Dart library for HTML5 parsing
-  html: ^0.15.0
+  html: ^0.15.3
+
+  # For app_theme.dart
+  google_fonts: ^6.2.1
+
+  # Determine if it is HMS device
+  google_api_availability: ^5.0.0
+
+  # From Dart Team: A composable, Future-based library for making HTTP requests.
+  http: '>=0.13.6 <2.0.0'
+
+  # From dart.dev: Platform independent access to information about the current operating system.
+  os_detect: ^2.0.1
+
+  # From Flutter Community: Provides an API for querying information about an application package.
+  package_info_plus: '>=4.0.1 <9.0.0'
+
+  # From Flutter Team: Wraps NSUserDefaults (on iOS) and SharedPreferences (on Android).
+  shared_preferences: '>=2.1.1 <3.0.0'
+
+  # From Flutter Team: A Flutter plugin for launching a URL in the mobile platform.
+  url_launcher: ^6.1.11
+
+  # A dart library for comparing and incrementing version numbers with Semantic Versioning.
+  version: ^3.0.2
+
+  # A dart library for parsing, traversing, querying, transforming and building XML documents.
+  xml: ^6.3.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
   # From Dart Team: Mock library for Dart inspired by Mockito.
-  mockito: ^5.0.0
-  flutter_lints: ^1.0.4
+  mockito: ^5.4.3
+  flutter_lints: ^2.0.3
+  go_router: ^7.1.1
 
 flutter:
