@@ -2,6 +2,8 @@
  * Copyright (c) 2020-2022 Larry Aasen. All rights reserved.
  */
 
+import 'dart:ui';
+
 import 'package:flutter/material.dart';
 import 'package:upgrader/src/message_constants.dart';
 
@@ -45,7 +47,9 @@ enum UpgraderMessage {
 ///   String get buttonTitleIgnore => 'My Ignore';
 /// }
 ///
-/// UpgradeAlert(messages: MyUpgraderMessages());
+/// final upgrader = Upgrader(messages: MyUpgraderMessages());
+/// ...
+/// UpgradeAlert(upgrader: upgrader);
 /// ```
 ///
 class UpgraderMessages {
@@ -87,13 +91,11 @@ class UpgraderMessages {
     Locale? locale;
     if (context != null) {
       locale = Localizations.maybeLocaleOf(context);
-    } else {
-      // Get the system locale
-      locale = ambiguate(WidgetsBinding.instance)!.window.locale;
     }
-    final code = locale == null || locale.languageCode.isEmpty
-        ? 'en'
-        : locale.languageCode;
+    // Get the system locale
+    locale ??= PlatformDispatcher.instance.locale;
+
+    final code = locale.languageCode.isEmpty ? 'en' : locale.languageCode;
     return code;
   }
 
@@ -112,6 +114,138 @@ class UpgraderMessages {
     switch (languageCode) {
       case 'ms':
         message = messageBody['ms']!;
+        break;
+      case 'ar':
+        message =
+            'نسخة جديدة من {{appName}} متوفرة! النسخة {{currentAppStoreVersion}} متوفرة الآن, أنت تستخدم النسخة {{currentInstalledVersion}}.';
+        break;
+      case 'bn':
+        message =
+            '{{appName}} এর একটি নতুন সংস্করণ {{currentAppStoreVersion}} পাওয়া যাচ্ছে। আপনার অ্যাপলিকেশনের সংস্করণ হচ্ছে {{currentInstalledVersion}}।';
+        break;
+      case 'da':
+        message =
+            'En ny version af {{appName}} er tilgængelig! Version {{currentAppStoreVersion}} er nu tilgængelig - du har {{currentInstalledVersion}}.';
+        break;
+      case 'el':
+        message =
+            'Μια νέα έκδοση του {{appName}} είναι διαθέσιμη! Η έκδοση {{currentAppStoreVersion}} είναι διαθέσιμη-έχετε την {{currentInstalledVersion}}.';
+        break;
+      case 'es':
+        message =
+            '¡Una nueva versión de {{appName}} está disponible! La versión {{currentAppStoreVersion}} ya está disponible-usted tiene {{currentInstalledVersion}}.';
+        break;
+      case 'fa':
+        message =
+            'نسخه‌ی جدیدی از {{appName}} موجود است! نسخه‌ی {{currentAppStoreVersion}} در دسترس است ولی شما همچنان از نسخه‌ی {{currentInstalledVersion}} استفاده می‌کنید.';
+        break;
+      case 'fil':
+        message =
+            'May bagong bersyon ang {{appName}} na pwede nang magamit! Ang bersyong {{currentAppStoreVersion}} ay pwede nang magamit. Ikaw ay kasalukuyang gumagamit ng bersyong {{currentInstalledVersion}}.';
+        break;
+      case 'fr':
+        message =
+            'Une nouvelle version de {{appName}} est disponible ! La version {{currentAppStoreVersion}} est maintenant disponible, vous avez la version {{currentInstalledVersion}}.';
+        break;
+      case 'de':
+        message =
+            'Eine neue Version von {{appName}} ist verfügbar! Die Version {{currentAppStoreVersion}} ist verfügbar, installiert ist die Version {{currentInstalledVersion}}.';
+        break;
+      case 'he':
+        message =
+            'גרסה חדשה של {{appName}} קיימת! גרסה {{currentAppStoreVersion}} ניתנת להורדה-יש לך גרסה {{currentInstalledVersion}}.';
+        break;
+      case 'hi':
+        message =
+            '{app name} का एक नया संस्करण उपलब्ध है। संस्करण {{currentAppStoreVersion}} अब उपलब्ध है-आपके पास है {{currentInstalledVersion}}.';
+        break;
+      case 'ht':
+        message =
+            'Yon nouvo vèsyon {{appName}} disponib! Vèsyon {{currentAppStoreVersion}} disponib, ou gen vèsyon {{currentInstalledVersion}}.';
+        break;
+      case 'hu':
+        message =
+            'Új verzió érhető el az alkalmazásból {{appName}} ! Az elérhető új verzió: {{currentAppStoreVersion}} - a jelenlegi verzió: {{currentInstalledVersion}}.';
+        break;
+      case 'id':
+        message =
+            'Versi terbaru dari {{appName}} tersedia! Versi terbaru saat ini adalah {{currentAppStoreVersion}} - versi anda saat ini adalah {{currentInstalledVersion}}.';
+        break;
+      case 'it':
+        message =
+            'Una nuova versione di {{appName}} è disponibile! La versione {{currentAppStoreVersion}} è ora disponibile, voi avete {{currentInstalledVersion}}.';
+        break;
+      case 'ja':
+        message =
+            '現在のバージョンは、{{currentInstalledVersion}}です。{{appName}}の最新バージョン({{currentAppStoreVersion}})があります。';
+        break;
+      case 'kk':
+        message =
+            '{{appName}} қосымша жаңа нұсқасын жүктеп алыңыз! Жаңа нұсқасы: {{currentAppStoreVersion}}, қазіргі нұсқасы: {{currentInstalledVersion}}';
+        break;
+      case 'km':
+        message =
+            'មានការអាប់ដេតថ្មីកម្មវិធី {{appName}} ហើយ! កំណែអាប់ដែត {{currentAppStoreVersion}} គឺអាចប្រើប្រាប់បានជំនួស {{currentInstalledVersion}} បានហើយ។';
+        break;
+      case 'ko':
+        message =
+            '{{appName}}이 새 버전으로 업데이트되었습니다! 최신 버전 {{currentAppStoreVersion}}으로 업그레이드 가능합니다 - 현재 버전 {{currentInstalledVersion}}.';
+        break;
+      case 'ku':
+        message =
+            'وەشانی نوێی {{appName}} بەردەستە! وەشانی {{currentAppStoreVersion}} بەردەستە- تۆ وەشانی {{currentInstalledVersion}} دابەزاندوە.';
+        break;
+      case 'lt':
+        message =
+            'Išleista nauja programos {{appName}} versija! Versija {{currentAppStoreVersion}} yra prieinama, jūs turite {{currentInstalledVersion}}.';
+        break;
+      case 'mn':
+        message =
+            '{{appName}}-н шинэ хувилбар бэлэн боллоо! Таны одоогийн ашиглаж буй хувилбар {{currentInstalledVersion}} - Шинээр бэлэн болсон хувилбар нь {{currentAppStoreVersion}} юм .';
+        break;
+      case 'nb':
+        message =
+            'En ny versjon av {{appName}} er tilgjengelig! {{currentAppStoreVersion}} er nå tilgjengelig - du har {{currentInstalledVersion}}.';
+        break;
+      case 'nl':
+        message =
+            'Er is een nieuwe versie van {{appName}} beschikbaar! De nieuwe versie is {{currentAppStoreVersion}}, je gebruikt nu versie {{currentInstalledVersion}}.';
+        break;
+      case 'pt':
+        message =
+            'Há uma nova versão do {{appName}} disponível! A versão {{currentAppStoreVersion}} já está disponível, você tem a {{currentInstalledVersion}}.';
+        break;
+      case 'pl':
+        message =
+            'Nowa wersja {{appName}} jest dostępna! Wersja {{currentAppStoreVersion}} jest dostępna, Ty masz {{currentInstalledVersion}}.';
+        break;
+      case 'ru':
+        message =
+            'Доступна новая версия приложения {{appName}}! Новая версия: {{currentAppStoreVersion}}, текущая версия: {{currentInstalledVersion}}.';
+        break;
+      case 'sv':
+        message =
+            'En ny version av {{appName}} är tillgänglig! Version {{currentAppStoreVersion}} är tillgänglig - du har {{currentInstalledVersion}}.';
+        break;
+      case 'ta':
+        message =
+            '{{appName}}-ன் புதிய பதிப்பு {{currentAppStoreVersion}} இப்போது கிடைக்கிறது! உங்களிடம் {{currentInstalledVersion}} உள்ளது.';
+        break;
+      case 'te':
+        message =
+            '{{appName}} యాప్ యొక్క కొత్త వెర్షన్ అందుబాటులో ఉంది. వెర్షన్ {{currentAppStoreVersion}} అందుబాటులో ఉంది కానీ మీ దగ్గర {{currentInstalledVersion}} ఉంది.';
+        break;
+      case 'tr':
+        message =
+            '{{appName}} uygulamanızın yeni bir versiyonu mevcut! Versiyon {{currentAppStoreVersion}} şu anda erişilebilir, mevcut sürümünüz {{currentInstalledVersion}}.';
+        break;
+      case 'uk':
+        message =
+            'Доступна нова версія додатка {{appName}}! Нова версія: {{currentAppStoreVersion}}, поточна версія: {{currentInstalledVersion}}.';
+        break;
+      case 'vi':
+        message =
+            'Đã có phiên bản mới của {{appName}}. Phiên bản {{currentAppStoreVersion}} đã sẵn sàng, bạn đang dùng {{currentInstalledVersion}}.';
         break;
       case 'zh':
         message = messageBody['zh']!;
@@ -133,6 +267,105 @@ class UpgraderMessages {
       case 'ms':
         message = messageButtonTitleIgnore['ms']!;
         break;
+      case 'ar':
+        message = 'تجاهل';
+        break;
+      case 'bn':
+        message = 'বাতিল';
+        break;
+      case 'da':
+        message = 'IGNORER';
+        break;
+      case 'el':
+        message = 'ΑΓΝΟΗΣTΕ';
+        break;
+      case 'es':
+        message = 'IGNORAR';
+        break;
+      case 'fa':
+        message = 'ردکردن';
+        break;
+      case 'fil':
+        message = 'HUWAG PANSININ';
+        break;
+      case 'fr':
+        message = 'IGNORER';
+        break;
+      case 'de':
+        message = 'IGNORIEREN';
+        break;
+      case 'he':
+        message = 'התעלם';
+        break;
+      case 'hi':
+        message = 'नज़रअंदाज़ करना';
+        break;
+      case 'ht':
+        message = 'INYORE';
+        break;
+      case 'hu':
+        message = 'KIHAGYOM';
+        break;
+      case 'id':
+        message = 'ABAIKAN';
+        break;
+      case 'it':
+        message = 'IGNORA';
+        break;
+      case 'ja':
+        message = '今はしない';
+        break;
+      case 'kk':
+        message = 'ЖОҚ';
+        break;
+      case 'km':
+        message = 'មិនអើពើ';
+        break;
+      case 'ko':
+        message = '무시';
+        break;
+      case 'ku':
+        message = 'پشتگوێخستن';
+        break;
+      case 'lt':
+        message = 'IGNORUOTI';
+        break;
+      case 'mn':
+        message = 'Татгалзах';
+        break;
+      case 'nb':
+        message = 'IGNORER';
+        break;
+      case 'nl':
+        message = 'NEGEREN';
+        break;
+      case 'pt':
+        message = 'IGNORAR';
+        break;
+      case 'pl':
+        message = 'IGNORUJ';
+        break;
+      case 'ru':
+        message = 'НЕТ';
+        break;
+      case 'sv':
+        message = 'AVBRYT';
+        break;
+      case 'ta':
+        message = 'புறக்கணி';
+        break;
+      case 'te':
+        message = 'తిరస్కరించండి';
+        break;
+      case 'tr':
+        message = 'YOKSAY';
+        break;
+      case 'uk':
+        message = 'НІ';
+        break;
+      case 'vi':
+        message = 'BỎ QUA';
+        break;
       case 'zh':
         message = messageButtonTitleIgnore['zh']!;
         break;
@@ -152,6 +385,108 @@ class UpgraderMessages {
     switch (languageCode) {
       case 'ms':
         message = messageButtonTitleLater['ms']!;
+        break;
+      case 'ar':
+        //only minor change here, removing the character at the left top of the arabic word (لاحقاً)
+        // before: message = 'لاحقاً';
+        //now:
+        message = 'لاحقا';
+        break;
+      case 'bn':
+        message = 'পরে';
+        break;
+      case 'da':
+        message = 'SENERE';
+        break;
+      case 'el':
+        message = 'ΑΡΓΟΤΕΡΑ';
+        break;
+      case 'es':
+        message = 'MÁS TARDE';
+        break;
+      case 'fa':
+        message = 'بعدا';
+        break;
+      case 'fil':
+        message = 'MAMAYA';
+        break;
+      case 'fr':
+        message = 'PLUS TARD';
+        break;
+      case 'de':
+        message = 'SPÄTER';
+        break;
+      case 'he':
+        message = 'אחר-כך';
+        break;
+      case 'hi':
+        message = 'बाद में';
+        break;
+      case 'ht':
+        message = 'PITA';
+        break;
+      case 'hu':
+        message = 'KÉSŐBB';
+        break;
+      case 'id':
+        message = 'NANTI';
+        break;
+      case 'it':
+        message = 'DOPO';
+        break;
+      case 'ja':
+        message = '後で通知';
+        break;
+      case 'kk':
+        message = 'КЕЙІН';
+        break;
+      case 'km':
+        message = 'ពេលក្រោយ';
+        break;
+      case 'ko':
+        message = '나중에';
+        break;
+      case 'ku':
+        message = 'دواتر';
+        break;
+      case 'lt':
+        message = 'ATNAUJINTI VĖLIAU';
+        break;
+      case 'mn':
+        message = 'Дараа суулгах';
+        break;
+      case 'nb':
+        message = 'SENERE';
+        break;
+      case 'nl':
+        message = 'LATER';
+        break;
+      case 'pt':
+        message = 'MAIS TARDE';
+        break;
+      case 'pl':
+        message = 'PÓŹNIEJ';
+        break;
+      case 'ru':
+        message = 'ПОЗЖЕ';
+        break;
+      case 'sv':
+        message = 'SENARE';
+        break;
+      case 'ta':
+        message = 'பிறகு';
+        break;
+      case 'te':
+        message = 'తరువాత';
+        break;
+      case 'tr':
+        message = 'SONRA';
+        break;
+      case 'uk':
+        message = 'ПІЗНІШЕ';
+        break;
+      case 'vi':
+        message = 'ĐỂ SAU';
         break;
       case 'zh':
         message = messageButtonTitleLater['zh']!;
@@ -173,6 +508,105 @@ class UpgraderMessages {
       case 'ms':
         message = messageButtonTitleUpdate['ms']!;
         break;
+      case 'ar':
+        message = 'حدث الآن';
+        break;
+      case 'bn':
+        message = 'এখন আপডেট করুন';
+        break;
+      case 'da':
+        message = 'OPDATER NU';
+        break;
+      case 'el':
+        message = 'ΕΝΗΜΕΡΩΣΗ';
+        break;
+      case 'es':
+        message = 'ACTUALIZAR';
+        break;
+      case 'fa':
+        message = 'بروزرسانی';
+        break;
+      case 'fil':
+        message = 'I-UPDATE NA NGAYON';
+        break;
+      case 'fr':
+        message = 'MAINTENANT';
+        break;
+      case 'de':
+        message = 'AKTUALISIEREN';
+        break;
+      case 'he':
+        message = 'עדכן';
+        break;
+      case 'hi':
+        message = 'अभी अद्यतन करें';
+        break;
+      case 'ht':
+        message = 'MIZAJOU KOUNYE A';
+        break;
+      case 'hu':
+        message = 'FRISSÍTSE MOST';
+        break;
+      case 'id':
+        message = 'PERBARUI SEKARANG';
+        break;
+      case 'it':
+        message = 'AGGIORNA ORA';
+        break;
+      case 'ja':
+        message = 'アップデート';
+        break;
+      case 'kk':
+        message = 'ЖАҢАРТУ';
+        break;
+      case 'km':
+        message = 'អាប់ដេតឥឡូវនេះ';
+        break;
+      case 'ko':
+        message = '지금 업데이트';
+        break;
+      case 'ku':
+        message = 'نوێکردنەوە';
+        break;
+      case 'lt':
+        message = 'ATNAUJINTI DABAR';
+        break;
+      case 'mn':
+        message = 'Шинэчлэх';
+        break;
+      case 'nb':
+        message = 'OPPDATER NÅ';
+        break;
+      case 'nl':
+        message = 'NU UPDATEN';
+        break;
+      case 'pt':
+        message = 'ATUALIZAR';
+        break;
+      case 'pl':
+        message = 'AKTUALIZUJ';
+        break;
+      case 'ru':
+        message = 'ОБНОВИТЬ';
+        break;
+      case 'sv':
+        message = 'UPPDATERA NU';
+        break;
+      case 'ta':
+        message = 'இப்பொழுது புதுப்பிக்கவும்';
+        break;
+      case 'te':
+        message = 'అప్‌డేట్‌ చేయండి';
+        break;
+      case 'tr':
+        message = 'ŞİMDİ GÜNCELLE';
+        break;
+      case 'uk':
+        message = 'ОНОВИТИ';
+        break;
+      case 'vi':
+        message = 'CẬP NHẬT';
+        break;
       case 'zh':
         message = messageButtonTitleUpdate['zh']!;
         break;
@@ -192,6 +626,105 @@ class UpgraderMessages {
     switch (languageCode) {
       case 'ms':
         message = messagePrompt['ms']!;
+        break;
+      case 'ar':
+        message = 'هل تفضل أن يتم التحديث الآن';
+        break;
+      case 'bn':
+        message = 'আপনি কি এখনই এটি আপডেট করতে চান?';
+        break;
+      case 'da':
+        message = 'Vil du opdatere nu?';
+        break;
+      case 'el':
+        message = 'Θέλετε να κάνετε την ενημέρωση τώρα;';
+        break;
+      case 'es':
+        message = '¿Le gustaría actualizar ahora?';
+        break;
+      case 'fa':
+        message = 'آیا بروزرسانی می‌کنید؟';
+        break;
+      case 'fil':
+        message = 'Gusto mo bang i-update ito ngayon?';
+        break;
+      case 'fr':
+        message = 'Voulez-vous mettre à jour maintenant ?';
+        break;
+      case 'de':
+        message = 'Möchtest du jetzt aktualisieren?';
+        break;
+      case 'he':
+        message = 'האם תרצה לעדכן עכשיו?';
+        break;
+      case 'hi':
+        message = 'क्या आप इसे अभी अद्यतन करना चाहेंगे?';
+        break;
+      case 'ht':
+        message = 'Èske ou vle mete aplikasyon an ajou kounye a?';
+        break;
+      case 'hu':
+        message = 'Akarja most frissíteni?';
+        break;
+      case 'id':
+        message = 'Apakah Anda ingin memperbaruinya sekarang?';
+        break;
+      case 'it':
+        message = 'Ti piacerebbe aggiornare ora?';
+        break;
+      case 'ja':
+        message = '今すぐアップデートしますか?';
+        break;
+      case 'kk':
+        message = 'Қазір жаңартқыңыз келе ме?';
+        break;
+      case 'km':
+        message = 'តើអ្នកចង់អាប់ដេតវាឥឡូវនេះទេ?';
+        break;
+      case 'ko':
+        message = '지금 업데이트를 시작하시겠습니까?';
+        break;
+      case 'ku':
+        message = 'دەتەوێت ئێستا نوێی بکەیەوە؟';
+        break;
+      case 'lt':
+        message = 'Ar norite atnaujinti dabar?';
+        break;
+      case 'mn':
+        message = 'Та одоо шинэчлэлтийг татаж авах уу?';
+        break;
+      case 'nb':
+        message = 'Ønsker du å oppdatere nå?';
+        break;
+      case 'nl':
+        message = 'Wil je de app nu updaten?';
+        break;
+      case 'pt':
+        message = 'Você quer atualizar agora?';
+        break;
+      case 'pl':
+        message = 'Czy chciałbyś zaktualizować teraz?';
+        break;
+      case 'ru':
+        message = 'Хотите обновить сейчас?';
+        break;
+      case 'sv':
+        message = 'Vill du uppdatera nu?';
+        break;
+      case 'ta':
+        message = 'இப்போது புதுப்பிக்க விரும்புகிறீர்களா?';
+        break;
+      case 'te':
+        message = 'మీరు దీన్ని ఇప్పుడే అప్‌డేట్ చేయాలనుకుంటున్నారా?';
+        break;
+      case 'tr':
+        message = 'Şimdi güncellemek ister misiniz?';
+        break;
+      case 'uk':
+        message = 'Бажаєте оновити зараз?';
+        break;
+      case 'vi':
+        message = 'Bạn có muốn cập nhật ứng dụng?';
         break;
       case 'zh':
         message = messagePrompt['zh']!;
@@ -213,6 +746,70 @@ class UpgraderMessages {
       case 'ms':
         message = messageReleaseNotes['ms']!;
         break;
+      case 'ar':
+        message = 'تفاصيل الاصدار';
+        break;
+      case 'da':
+        message = 'Udgivelsesnoter';
+        break;
+      case 'de':
+        message = 'Versionshinweise';
+        break;
+      case 'es':
+        message = 'Notas De Lanzamiento';
+        break;
+      case 'fr':
+        message = 'Notes de version';
+        break;
+      case 'he':
+        message = 'חדש בגרסה';
+        break;
+      case 'hi':
+        message = 'रिहाई टिप्पणी';
+        break;
+      case 'id':
+        message = 'Catatan Rilis';
+        break;
+      case 'it':
+        message = 'Note di rilascio';
+        break;
+      case 'ja':
+        message = 'リリースノート';
+        break;
+      case 'ku':
+        message = 'تیبینەکانی وەشان';
+        break;
+      case 'pt':
+        message = 'Novidades';
+        break;
+      case 'ru':
+        message = 'Информация о выпуске';
+        break;
+      case 'te':
+        message = 'విడుదల గమనికలు';
+        break;
+      case 'tr':
+        message = 'Yayın Notları';
+        break;
+
+      case 'bn':
+      case 'el':
+      case 'fa':
+      case 'fil':
+      case 'ht':
+      case 'hu':
+      case 'kk':
+      case 'km':
+      case 'ko':
+      case 'lt':
+      case 'mn':
+      case 'nb':
+      case 'nl':
+      case 'pl':
+      case 'sv':
+      case 'ta':
+      case 'uk':
+      case 'vi':
       case 'zh':
         message = messageReleaseNotes['zh']!;
         break;
@@ -232,6 +829,105 @@ class UpgraderMessages {
     switch (languageCode) {
       case 'ms':
         message = messageTitle['ms']!;
+        break;
+      case 'ar':
+        message = 'هل تريد تحديث التطبيق؟';
+        break;
+      case 'bn':
+        message = 'আপডেট অ্যাপ্লিকেশন?';
+        break;
+      case 'da':
+        message = 'Opdater App?';
+        break;
+      case 'el':
+        message = 'Ενημέρωση εφαρμογής;';
+        break;
+      case 'es':
+        message = '¿Actualizar la aplicación?';
+        break;
+      case 'fa':
+        message = 'نسخه‌ی جدید';
+        break;
+      case 'fil':
+        message = 'I-update ang app?';
+        break;
+      case 'fr':
+        message = 'Mettre à jour l\'application ?';
+        break;
+      case 'de':
+        message = 'App aktualisieren?';
+        break;
+      case 'he':
+        message = 'לעדכן יישומון?';
+        break;
+      case 'hi':
+        message = 'अद्यतन ऐप?';
+        break;
+      case 'ht':
+        message = 'Mete app la ajou?';
+        break;
+      case 'hu':
+        message = 'FrissÍtés?';
+        break;
+      case 'id':
+        message = 'Perbarui Aplikasi?';
+        break;
+      case 'it':
+        message = 'Aggiornare l\'applicazione?';
+        break;
+      case 'ja':
+        message = 'アプリのアップデート';
+        break;
+      case 'kk':
+        message = 'Жаңарту керек пе?';
+        break;
+      case 'km':
+        message = 'អាប់ដេតកម្មវិធីទេ?';
+        break;
+      case 'ko':
+        message = '앱을 업데이트하시겠습니까?';
+        break;
+      case 'ku':
+        message = 'نوێکردنەوەی ئەپ؟';
+        break;
+      case 'lt':
+        message = 'Atnaujinti programą?';
+        break;
+      case 'mn':
+        message = 'Та шинэчлэлт хийх үү?';
+        break;
+      case 'nb':
+        message = 'Oppdater app?';
+        break;
+      case 'nl':
+        message = 'App updaten?';
+        break;
+      case 'pt':
+        message = 'Atualizar aplicação?';
+        break;
+      case 'pl':
+        message = 'Czy zaktualizować aplikację?';
+        break;
+      case 'ru':
+        message = 'Обновить?';
+        break;
+      case 'sv':
+        message = 'Uppdatera App?';
+        break;
+      case 'ta':
+        message = 'செயலியை புதுப்பிக்கவா?';
+        break;
+      case 'te':
+        message = 'యాప్‌ని అప్‌డేట్‌ చేయాలా?';
+        break;
+      case 'tr':
+        message = 'Uygulamayı Güncelle?';
+        break;
+      case 'uk':
+        message = 'Оновити?';
+        break;
+      case 'vi':
+        message = 'Cập nhật ứng dụng?';
         break;
       case 'zh':
         message = messageTitle['zh']!;
