@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-2021 Larry Aasen. All rights reserved.
+ * Copyright (c) 2018-2024 Larry Aasen. All rights reserved.
  */
 
 library upgrader;
@@ -9,8 +9,12 @@ export 'src/appcast.dart';
 export 'src/itunes_search_api.dart';
 export 'src/play_store_search_api.dart';
 export 'src/upgrade_alert.dart';
-export 'src/upgrade_base.dart';
 export 'src/upgrade_card.dart';
-export 'src/upgrade_io.dart';
+export 'src/upgrade_device.dart';
 export 'src/upgrade_messages.dart';
+export 'src/upgrade_os.dart';
+export 'src/upgrade_state.dart';
+export 'src/upgrade_store_controller.dart';
 export 'src/upgrader.dart';
+export 'src/app_update_dialog.dart';
+export 'src/upgrader_version_info.dart';
