@@ -10,8 +10,10 @@ import 'package:version/version.dart';
 import 'mock_play_store_client.dart';
 
 /// Helper method
-String? pmav(Document response, {String tagName = 'mav'}) {
-  final mav = PlayStoreResults.minAppVersion(response, tagName: tagName);
+String? pmav(Document response,
+    {String tagRES = r'\[\:mav\:[\s]*(?<version>[^\s]+)[\s]*\]'}) {
+  final mav =
+      PlayStoreSearchAPI().minAppVersion(response, tagRegExpSource: tagRES);
   return mav?.toString();
 }
 
@@ -25,33 +27,39 @@ void main() {
     expect(Version.parse('1.2.3+1').toString(), '1.2.3+1');
     expect(Version.parse('0.0.0').toString(), '0.0.0');
     expect(Version.parse('0.0.0+1').toString(), '0.0.0+1');
+
+    final version1 = Version.parse('1.2.3+1');
+    final version2 = Version.parse('1.2.3+2');
+    expect(version1 == version2, isTrue);
   }, skip: false);
 
   test('testing PlayStoreSearchAPI properties', () async {
     final playStore = PlayStoreSearchAPI();
-    expect(playStore.debugEnabled, equals(false));
-    playStore.debugEnabled = true;
-    expect(playStore.debugEnabled, equals(true));
+    expect(playStore.debugLogging, equals(false));
+    playStore.debugLogging = true;
+    expect(playStore.debugLogging, equals(true));
     expect(playStore.playStorePrefixURL.length, greaterThan(0));
 
     expect(
         playStore.lookupURLById('com.kotoko.express'),
         startsWith(
-            'https://play.google.com/store/apps/details?id=com.kotoko.express&gl=US&_cb='));
+            'https://play.google.com/store/apps/details?id=com.kotoko.express&gl=US&hl=en&_cb='));
   }, skip: false);
 
   test('testing lookupById', () async {
-    final client = await MockPlayStoreSearchClient.setupMockClient();
-    final playStore = PlayStoreSearchAPI(client: client);
+    final client = await MockPlayStoreSearchClient.setupMockClient(
+        verifyHeaders: {'header1': 'value1'});
+    final playStore = PlayStoreSearchAPI(
+        client: client, clientHeaders: {'header1': 'value1'});
     expect(() async => await playStore.lookupById(''), throwsAssertionError);
 
     final response = await playStore.lookupById('com.kotoko.express');
     expect(response, isNotNull);
     expect(response, isInstanceOf<Document>());
 
-    expect(PlayStoreResults.releaseNotes(response!),
-        'Minor updates and improvements.');
-    expect(PlayStoreResults.version(response), '1.23.0');
+    expect(
+        playStore.releaseNotes(response!), 'Minor updates and improvements.');
+    expect(playStore.version(response), '1.23.0');
 
     expect(await playStore.lookupById('com.not.a.valid.application'), isNull);
 
@@ -73,26 +81,30 @@ void main() {
     expect(() => playStore.lookupURLById(''), throwsAssertionError);
     expect(
         playStore.lookupURLById('com.testing.test1')!.startsWith(
-            'https://play.google.com/store/apps/details?id=com.testing.test1&gl=US&_cb=16'),
+            'https://play.google.com/store/apps/details?id=com.testing.test1&gl=US&hl=en&_cb=17'),
         equals(true));
     expect(
         playStore.lookupURLById('com.testing.test1', country: null)!.startsWith(
-            'https://play.google.com/store/apps/details?id=com.testing.test1&_cb=16'),
+            'https://play.google.com/store/apps/details?id=com.testing.test1&hl=en&_cb=17'),
         equals(true));
     expect(
         playStore.lookupURLById('com.testing.test1', country: '')!.startsWith(
-            'https://play.google.com/store/apps/details?id=com.testing.test1&_cb=16'),
+            'https://play.google.com/store/apps/details?id=com.testing.test1&hl=en&_cb=17'),
         equals(true));
     expect(
         playStore.lookupURLById('com.testing.test1', country: 'IN')!.startsWith(
-            'https://play.google.com/store/apps/details?id=com.testing.test1&gl=IN&_cb=16'),
+            'https://play.google.com/store/apps/details?id=com.testing.test1&gl=IN&hl=en&_cb=17'),
+        equals(true));
+    expect(
+        playStore.lookupURLById('com.testing.test1', language: 'es')!.startsWith(
+            'https://play.google.com/store/apps/details?id=com.testing.test1&gl=US&hl=es&_cb=17'),
         equals(true));
     expect(
         playStore
             .lookupURLById('com.testing.test1',
                 country: 'IN', useCacheBuster: false)!
             .startsWith(
-                'https://play.google.com/store/apps/details?id=com.testing.test1&gl=IN'),
+                'https://play.google.com/store/apps/details?id=com.testing.test1&gl=IN&hl=en'),
         equals(true));
   }, skip: false);
 
@@ -104,13 +116,53 @@ void main() {
     expect(response, isNotNull);
     expect(response, isInstanceOf<Document>());
 
-    expect(PlayStoreResults.releaseNotes(response!),
-        'Minor updates and improvements.');
-    expect(PlayStoreResults.version(response), '2.3.0');
-    expect(PlayStoreResults.description(response)?.length, greaterThan(10));
-    expect(pmav(response), '2.0.0');
+    expect(
+        playStore.releaseNotes(response!), 'Minor updates and improvements.');
+    expect(playStore.version(response), '2.3.0');
+    expect(playStore.description(response)?.length, greaterThan(10));
+    expect(
+        pmav(response,
+            tagRES:
+                r'\[\Minimum supported app version\:[\s]*(?<version>[^\s]+)[\s]*\]'),
+        '2.0.0');
 
     expect(await playStore.lookupById('com.not.a.valid.application'), isNull);
+  }, skip: false);
+
+  test(
+      'testing lookupById with redesignedVersion title with special characters',
+      () async {
+    final client = await MockPlayStoreSearchClient.setupMockClient();
+    final playStore = PlayStoreSearchAPI(client: client);
+
+    final response = await playStore.lookupById('com.testing.test8');
+    expect(response, isNotNull);
+    expect(response, isInstanceOf<Document>());
+
+    expect(
+        playStore.releaseNotes(response!), 'Minor updates and improvements.');
+    expect(playStore.version(response), '2.3.0');
+    expect(playStore.description(response)?.length, greaterThan(10));
+    expect(
+        pmav(response,
+            tagRES:
+                r'\[\Minimum supported app version\:[\s]*(?<version>[^\s]+)[\s]*\]'),
+        '2.0.0');
+
+    expect(await playStore.lookupById('com.not.a.valid.application'), isNull);
+  }, skip: false);
+
+  test('testing lookupById with invalid version', () async {
+    final client = await MockPlayStoreSearchClient.setupMockClient();
+    final playStore = PlayStoreSearchAPI(client: client);
+
+    final response = await playStore.lookupById('com.testing.test7');
+    expect(response, isNotNull);
+    expect(response, isInstanceOf<Document>());
+
+    expect(
+        playStore.releaseNotes(response!), 'Minor updates and improvements.');
+    expect(playStore.version(response), isNull);
   }, skip: false);
 
   test('testing release notes', () async {
@@ -121,10 +173,10 @@ void main() {
     expect(response, isNotNull);
     expect(response, isInstanceOf<Document>());
 
-    expect(PlayStoreResults.releaseNotes(response!),
-        'Minor updates and improvements.');
-    expect(PlayStoreResults.version(response), '2.0.2');
-    expect(PlayStoreResults.description(response)?.length, greaterThan(10));
+    expect(
+        playStore.releaseNotes(response!), 'Minor updates and improvements.');
+    expect(playStore.version(response), '2.0.2');
+    expect(playStore.description(response)?.length, greaterThan(10));
   }, skip: false);
 
   test('testing release notes <br>', () async {
@@ -135,16 +187,34 @@ void main() {
     expect(response, isNotNull);
     expect(response, isInstanceOf<Document>());
 
-    expect(PlayStoreResults.releaseNotes(response!),
+    expect(playStore.releaseNotes(response!),
         'Minor updates and improvements.\nAgain.\nAgain.');
-    expect(PlayStoreResults.version(response), '2.0.2');
-    expect(PlayStoreResults.description(response)?.length, greaterThan(10));
+    expect(playStore.version(response), '2.0.2');
+    expect(playStore.description(response)?.length, greaterThan(10));
   }, skip: false);
 
-  test('testing PlayStoreResults', () async {
-    expect(PlayStoreResults(), isNotNull);
-    expect(PlayStoreResults.releaseNotes(Document()), isNull);
-    expect(PlayStoreResults.version(Document()), isNull);
+  test('testing release notes <br> 2', () async {
+    final client = await MockPlayStoreSearchClient.setupMockClient();
+    final playStore = PlayStoreSearchAPI(client: client);
+
+    final response = await playStore.lookupById('com.testing.test5');
+    expect(response, isNotNull);
+    expect(response, isInstanceOf<Document>());
+
+    expect(playStore.releaseNotes(response!),
+        'Minor updates and improvements.\nAgain.\nAgain.');
+    expect(playStore.version(response), '2.0.2');
+    expect(playStore.description(response)?.length, greaterThan(10));
+  }, skip: false);
+
+  test('testing invalid store version', () async {
+    final client = await MockPlayStoreSearchClient.setupMockClient();
+    final playStore = PlayStoreSearchAPI(client: client);
+
+    final response = await playStore.lookupById('com.testing.test6');
+    expect(response, isNotNull);
+    expect(response, isInstanceOf<Document>());
+    expect(playStore.version(response!), isNull);
   }, skip: false);
 
   /// Helper method
@@ -168,7 +238,22 @@ void main() {
   }, skip: false);
 
   test('testing minAppVersion mav tag', () async {
-    expect(pmav(resDesc('test [:mav: 1.2.3]'), tagName: 'ddd'), isNull);
-    expect(pmav(resDesc('test [:ddd: 1.2.3]'), tagName: 'ddd'), '1.2.3');
+    expect(pmav(resDesc('test [:mav: 1.2.3]'), tagRES: 'ddd'), isNull);
+    expect(pmav(resDesc('test [:mav: a.b.c]')), isNull);
+    expect(
+        pmav(resDesc('test [:ddd: 1.2.3]'),
+            tagRES: r'\[\:ddd\:[\s]*(?<version>[^\s]+)[\s]*\]'),
+        '1.2.3');
+    expect(
+        pmav(resDesc('test [Minimum supported app version: 4.5.6+1]'),
+            tagRES:
+                r'\[\Minimum supported app version\:[\s]*(?<version>[^\s]+)[\s]*\]'),
+        '4.5.6+1');
+  });
+
+  test('testing special characters', () async {
+    const msg = 'McDonald\u0027s';
+    expect(msg.replaceAll(r"\u0027", "'"), 'McDonald\'s');
+    expect(msg.replaceAll(r"\u0027", '\''), 'McDonald\'s');
   });
 }
