@@ -1,22 +1,25 @@
 /*
- * Copyright (c) 2018-2022 Larry Aasen. All rights reserved.
+ * Copyright (c) 2018-2024 Larry Aasen. All rights reserved.
  */
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:go_router/go_router.dart';
+import 'package:http/src/client.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:upgrader/upgrader.dart';
+import 'package:version/version.dart';
 
+import 'appcast_test.dart';
 import 'fake_appcast.dart';
 import 'mock_itunes_client.dart';
 import 'mock_play_store_client.dart';
+import 'test_utils.dart';
 
-// TODO: Need an integration test that runs on Android and iOS.
-
-// Platform.operatingSystem can be "macos" or "linux" in a unit test.
-// defaultTargetPlatform is TargetPlatform.android in a unit test.
+// FYI: Platform.operatingSystem can be "macos" or "linux" in a unit test.
+// FYI: defaultTargetPlatform is TargetPlatform.android in a unit test.
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
@@ -32,7 +35,8 @@ void main() {
     return true;
   });
 
-  testWidgets('test Upgrader sharedInstance', (WidgetTester tester) async {
+  testWidgets('test Upgrader sharedInstance always returns same instance',
+      (WidgetTester tester) async {
     final upgrader1 = Upgrader.sharedInstance;
     expect(upgrader1, isNotNull);
     final upgrader2 = Upgrader.sharedInstance;
@@ -40,51 +44,183 @@ void main() {
     expect(upgrader1 == upgrader2, isTrue);
   }, skip: false);
 
+  testWidgets('test Upgrader multiple instances', (WidgetTester tester) async {
+    await tester.runAsync(() async {
+      final packageInfo = PackageInfo(
+          appName: 'Upgrader',
+          packageName: 'com.larryaasen.upgrader',
+          version: '1.9.9',
+          buildNumber: '400');
+
+      final client = MockITunesSearchClient.setupMockClient();
+      final upgrader = Upgrader(
+          upgraderOS: MockUpgraderOS(ios: true),
+          client: client,
+          debugLogging: true);
+
+      expect(tester.takeException(), null);
+      await tester.pumpAndSettle();
+      try {
+        expect(upgrader.appName(), 'Upgrader');
+      } catch (e) {
+        expect(e, Upgrader.notInitializedExceptionMessage);
+      }
+
+      upgrader.installPackageInfo(packageInfo: packageInfo);
+      expect(await upgrader.initialize(), isTrue);
+
+      final upgrader1 = Upgrader(
+          upgraderOS: MockUpgraderOS(ios: true),
+          client: client,
+          debugLogging: true);
+      upgrader1.installPackageInfo(packageInfo: packageInfo);
+      expect(await upgrader1.initialize(), isTrue);
+    });
+  });
+
+  // testWidgets('test Upgrader no package info', (WidgetTester tester) async {
+  //   await tester.runAsync(() async {
+  //     final client = MockITunesSearchClient.setupMockClient();
+  //     final upgrader = Upgrader(
+  //       upgraderOS: MockUpgraderOS(ios: true),
+  //       client: client,
+  //       debugLogging: true,
+  //     );
+
+  //     expect(tester.takeException(), null);
+  //     await tester.pumpAndSettle();
+  //     try {
+  //       expect(upgrader.appName(), 'Upgrader');
+  //     } catch (e) {
+  //       expect(e, Upgrader.notInitializedExceptionMessage);
+  //     }
+
+  //     expect(await upgrader.initialize(), isTrue);
+  //   });
+  // });
+
+  testWidgets('test Upgrader clearSavedSettings', (WidgetTester tester) async {
+    await Upgrader.clearSavedSettings();
+  }, skip: false);
+
+  testWidgets('test Upgrader stand alone', (WidgetTester tester) async {
+    final client = MockITunesSearchClient.setupMockClient();
+    final upgrader = Upgrader(
+        upgraderOS: MockUpgraderOS(ios: true),
+        client: client,
+        debugLogging: true);
+
+    upgrader.installPackageInfo(
+        packageInfo: PackageInfo(
+            appName: 'Upgrader',
+            packageName: 'com.larryaasen.upgrader',
+            version: '0.9.9',
+            buildNumber: '400'));
+
+    upgrader.initialize().then((value) {});
+
+    await tester.pumpAndSettle();
+
+    expect(upgrader.isUpdateAvailable(), true);
+    expect(upgrader.currentAppStoreVersion, '5.6.0');
+  });
+
   testWidgets('test Upgrader class', (WidgetTester tester) async {
-    final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader(platform: TargetPlatform.iOS, client: client);
+    await tester.runAsync(() async {
+      final client = MockITunesSearchClient.setupMockClient();
+      final upgrader = Upgrader(
+          upgraderOS: MockUpgraderOS(ios: true),
+          client: client,
+          debugLogging: true);
 
-    expect(tester.takeException(), null);
-    await tester.pumpAndSettle();
-    try {
+      expect(tester.takeException(), null);
+      await tester.pumpAndSettle();
+      try {
+        expect(upgrader.appName(), 'Upgrader');
+      } catch (e) {
+        expect(e, Upgrader.notInitializedExceptionMessage);
+      }
+
+      upgrader.installPackageInfo(
+          packageInfo: PackageInfo(
+              appName: 'Upgrader',
+              packageName: 'com.larryaasen.upgrader',
+              version: '1.9.9',
+              buildNumber: '400'));
+
+      expect(await upgrader.initialize(), isTrue);
+
+      // Calling initialize() a second time should do nothing
+      expect(await upgrader.initialize(), isTrue);
+
       expect(upgrader.appName(), 'Upgrader');
-    } catch (e) {
-      expect(e, upgrader.notInitializedExceptionMessage);
-    }
+      expect(upgrader.currentAppStoreVersion, '5.6.0');
+      expect(upgrader.currentInstalledVersion, '1.9.9');
+      expect(upgrader.isUpdateAvailable(), true);
 
-    upgrader.installPackageInfo(
-        packageInfo: PackageInfo(
-            appName: 'Upgrader',
-            packageName: 'com.larryaasen.upgrader',
-            version: '1.9.9',
-            buildNumber: '400'));
-    await upgrader.initialize();
+      upgrader.updateState(upgrader.state.copyWith(
+          versionInfo: UpgraderVersionInfo(appStoreVersion: Version(1, 2, 3))));
+      expect(upgrader.currentAppStoreVersion, '1.2.3');
+      expect(upgrader.isUpdateAvailable(), false);
 
-    // Calling initialize() a second time should do nothing
-    await upgrader.initialize();
+      upgrader.updateState(upgrader.state.copyWith(
+          versionInfo: UpgraderVersionInfo(appStoreVersion: Version(6, 2, 3))));
+      expect(upgrader.currentAppStoreVersion, '6.2.3');
+      expect(upgrader.isUpdateAvailable(), true);
 
-    expect(upgrader.appName(), 'Upgrader');
-    expect(upgrader.currentAppStoreVersion(), '5.6');
-    expect(upgrader.currentInstalledVersion(), '1.9.9');
-    expect(upgrader.isUpdateAvailable(), true);
+      upgrader.updateState(upgrader.state.copyWith(
+          versionInfo: UpgraderVersionInfo(appStoreVersion: Version(1, 1, 1))));
+      expect(upgrader.currentAppStoreVersion, '1.1.1');
+      expect(upgrader.isUpdateAvailable(), false);
 
-    upgrader.installAppStoreVersion('1.2.3');
-    expect(upgrader.currentAppStoreVersion(), '1.2.3');
-  }, skip: false);
+      await upgrader.didChangeAppLifecycleState(AppLifecycleState.resumed);
+      expect(upgrader.isUpdateAvailable(), true);
 
-  testWidgets('test installAppStoreListingURL', (WidgetTester tester) async {
-    final upgrader = Upgrader();
-    upgrader.installAppStoreListingURL(
-        'https://itunes.apple.com/us/app/google-maps-transit-food/id585027354?mt=8&uo=4');
+      upgrader.updateState(upgrader.state.copyWith(
+          versionInfo: UpgraderVersionInfo(appStoreVersion: Version(1, 1, 1))));
+      expect(upgrader.currentAppStoreVersion, '1.1.1');
+      expect(upgrader.isUpdateAvailable(), false);
 
-    expect(upgrader.currentAppStoreListingURL(),
-        'https://itunes.apple.com/us/app/google-maps-transit-food/id585027354?mt=8&uo=4');
-  }, skip: false);
+      upgrader.installPackageInfo(
+          packageInfo: PackageInfo(
+              appName: 'Upgrader',
+              packageName: 'com.larryaasen.upgrader.2',
+              version: '1.9.9',
+              buildNumber: '400'));
 
-  testWidgets('test UpgradeWidget', (WidgetTester tester) async {
+      await upgrader.didChangeAppLifecycleState(AppLifecycleState.resumed);
+      expect(upgrader.isUpdateAvailable(), true);
+      expect(upgrader.currentAppStoreVersion, '7.0.0');
+
+      upgrader.installPackageInfo(
+          packageInfo: PackageInfo(
+              appName: 'Upgrader',
+              packageName: 'com.larryaasen.upgrader.3',
+              version: '1.9.9',
+              buildNumber: '400'));
+
+      await upgrader.didChangeAppLifecycleState(AppLifecycleState.resumed);
+      expect(upgrader.isUpdateAvailable(), false);
+
+      upgrader.installPackageInfo(
+          packageInfo: PackageInfo(
+              appName: 'Upgrader',
+              packageName: 'com.larryaasen.upgrader.4',
+              version: '1.9.9',
+              buildNumber: '400'));
+
+      await upgrader.didChangeAppLifecycleState(AppLifecycleState.resumed);
+      expect(upgrader.currentAppStoreVersion, isNull);
+      expect(upgrader.isUpdateAvailable(), false);
+    });
+  });
+
+  testWidgets('test UpgradeAlert', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
     final upgrader = Upgrader(
-        platform: TargetPlatform.iOS, client: client, debugLogging: true);
+        upgraderOS: MockUpgraderOS(ios: true),
+        client: client,
+        debugLogging: true);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -92,76 +228,101 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
-
-    var called = false;
-    var notCalled = true;
-    upgrader.onUpdate = () {
-      called = true;
-      return true;
-    };
-    upgrader.onIgnore = () {
-      notCalled = false;
-      return true;
-    };
-    upgrader.onLater = () {
-      notCalled = false;
-      return true;
-    };
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     expect(upgrader.isUpdateAvailable(), true);
     expect(upgrader.isTooSoon(), false);
 
-    expect(upgrader.messages, isNotNull);
+    expect(upgrader.state.messages, isNull);
+    upgrader.updateState(upgrader.state.copyWith(messages: UpgraderMessages()));
+    upgrader.updateState(upgrader.state.copyWith(messages: UpgraderMessages()));
+    expect(upgrader.state.messages, isNotNull);
 
-    expect(upgrader.messages.buttonTitleIgnore, 'IGNORE');
-    expect(upgrader.messages.buttonTitleLater, 'LATER');
-    expect(upgrader.messages.buttonTitleUpdate, 'UPDATE NOW');
-    expect(upgrader.messages.releaseNotes, 'Release Notes');
+    expect(upgrader.state.messages?.buttonTitleIgnore, 'IGNORE');
+    expect(upgrader.state.messages?.buttonTitleLater, 'LATER');
+    expect(upgrader.state.messages?.buttonTitleUpdate, 'UPDATE NOW');
+    expect(upgrader.state.messages?.releaseNotes, 'Release Notes');
 
-    upgrader.messages = MyUpgraderMessages();
+    upgrader
+        .updateState(upgrader.state.copyWith(messages: MyUpgraderMessages()));
 
-    expect(upgrader.messages.buttonTitleIgnore, 'aaa');
-    expect(upgrader.messages.buttonTitleLater, 'bbb');
-    expect(upgrader.messages.buttonTitleUpdate, 'ccc');
-    expect(upgrader.messages.releaseNotes, 'ddd');
+    expect(upgrader.state.messages!.buttonTitleIgnore, 'aaa');
+    expect(upgrader.state.messages!.buttonTitleLater, 'bbb');
+    expect(upgrader.state.messages!.buttonTitleUpdate, 'ccc');
+    expect(upgrader.state.messages!.releaseNotes, 'ddd');
 
-    await tester.pumpWidget(_MyWidget(upgrader: upgrader));
+    var called = false;
+    var notCalled = true;
+
+    final dialogKey = GlobalKey(debugLabel: 'gloabl_upgrader_alert_dialog');
+    final upgradeAlert = wrapper(
+      UpgradeAlert(
+        dialogKey: dialogKey,
+        upgrader: upgrader,
+        onUpdate: () {
+          called = true;
+          return true;
+        },
+        onIgnore: () {
+          notCalled = false;
+          return true;
+        },
+        onLater: () {
+          notCalled = false;
+          return true;
+        },
+        child: const Center(child: Text('Upgrading')),
+      ),
+    );
+    await tester.pumpWidget(upgradeAlert);
 
     expect(find.text('Upgrader test'), findsOneWidget);
     expect(find.text('Upgrading'), findsOneWidget);
 
     // Pump the UI so the upgrader can display its dialog
     await tester.pumpAndSettle();
-    await tester.pumpAndSettle();
 
     expect(upgrader.isTooSoon(), true);
 
-    expect(find.text(upgrader.messages.title), findsOneWidget);
-    expect(find.text(upgrader.message()), findsOneWidget);
-    expect(find.text(upgrader.messages.releaseNotes), findsOneWidget);
+    expect(find.text(upgrader.state.messages!.title), findsOneWidget);
+    expect(find.text(upgrader.body(upgrader.state.messages!)), findsOneWidget);
+    expect(find.text(upgrader.state.messages!.releaseNotes), findsOneWidget);
     expect(find.text(upgrader.releaseNotes!), findsOneWidget);
-    expect(find.text(upgrader.messages.prompt), findsOneWidget);
+    expect(find.text(upgrader.state.messages!.prompt), findsOneWidget);
     expect(find.byType(TextButton), findsNWidgets(3));
-    expect(find.text(upgrader.messages.buttonTitleIgnore), findsOneWidget);
-    expect(find.text(upgrader.messages.buttonTitleLater), findsOneWidget);
-    expect(find.text(upgrader.messages.buttonTitleUpdate), findsOneWidget);
-    expect(find.text(upgrader.messages.releaseNotes), findsOneWidget);
+    expect(
+        find.text(upgrader.state.messages!.buttonTitleIgnore), findsOneWidget);
+    expect(
+        find.text(upgrader.state.messages!.buttonTitleLater), findsOneWidget);
+    expect(
+        find.text(upgrader.state.messages!.buttonTitleUpdate), findsOneWidget);
+    expect(find.text(upgrader.state.messages!.releaseNotes), findsOneWidget);
+    expect(find.byKey(dialogKey), findsOneWidget);
 
-    await tester.tap(find.text(upgrader.messages.buttonTitleUpdate));
+    await tester.tap(find.text(upgrader.state.messages!.buttonTitleUpdate));
     await tester.pumpAndSettle();
-    expect(find.text(upgrader.messages.buttonTitleIgnore), findsNothing);
-    expect(find.text(upgrader.messages.buttonTitleLater), findsNothing);
-    expect(find.text(upgrader.messages.buttonTitleUpdate), findsNothing);
-    expect(find.text(upgrader.messages.releaseNotes), findsNothing);
+    expect(find.text(upgrader.state.messages!.buttonTitleIgnore), findsNothing);
+    expect(find.text(upgrader.state.messages!.buttonTitleLater), findsNothing);
+    expect(find.text(upgrader.state.messages!.buttonTitleUpdate), findsNothing);
+    expect(find.text(upgrader.state.messages!.releaseNotes), findsNothing);
     expect(called, true);
     expect(notCalled, true);
+    // });
   }, skip: false);
 
-  testWidgets('test UpgradeWidget Cupertino', (WidgetTester tester) async {
+  testWidgets('test UpgradeAlert Cupertino', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
+
+    const cupertinoButtonTextStyle = TextStyle(
+      fontSize: 14,
+      color: Colors.green,
+    );
     final upgrader = Upgrader(
-        platform: TargetPlatform.iOS, client: client, debugLogging: true);
+      upgraderOS: MockUpgraderOS(ios: true),
+      client: client,
+      debugLogging: true,
+    );
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -169,71 +330,96 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
-
-    var called = false;
-    var notCalled = true;
-    upgrader.onUpdate = () {
-      called = true;
-      return true;
-    };
-    upgrader.onIgnore = () {
-      notCalled = false;
-      return true;
-    };
-    upgrader.onLater = () {
-      notCalled = false;
-      return true;
-    };
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     expect(upgrader.isUpdateAvailable(), true);
     expect(upgrader.isTooSoon(), false);
 
-    expect(upgrader.messages, isNotNull);
+    expect(upgrader.state.messages, isNull);
+    upgrader.updateState(upgrader.state.copyWith(messages: UpgraderMessages()));
+    expect(upgrader.state.messages, isNotNull);
 
-    expect(upgrader.messages.buttonTitleIgnore, 'IGNORE');
-    expect(upgrader.messages.buttonTitleLater, 'LATER');
-    expect(upgrader.messages.buttonTitleUpdate, 'UPDATE NOW');
+    expect(upgrader.state.messages!.buttonTitleIgnore, 'IGNORE');
+    expect(upgrader.state.messages!.buttonTitleLater, 'LATER');
+    expect(upgrader.state.messages!.buttonTitleUpdate, 'UPDATE NOW');
 
-    upgrader.messages = MyUpgraderMessages();
+    upgrader
+        .updateState(upgrader.state.copyWith(messages: MyUpgraderMessages()));
 
-    expect(upgrader.messages.buttonTitleIgnore, 'aaa');
-    expect(upgrader.messages.buttonTitleLater, 'bbb');
-    expect(upgrader.messages.buttonTitleUpdate, 'ccc');
-    upgrader.dialogStyle = UpgradeDialogStyle.cupertino;
+    expect(upgrader.state.messages!.buttonTitleIgnore, 'aaa');
+    expect(upgrader.state.messages!.buttonTitleLater, 'bbb');
+    expect(upgrader.state.messages!.buttonTitleUpdate, 'ccc');
 
-    await tester.pumpWidget(_MyWidget(upgrader: upgrader));
+    var called = false;
+    var notCalled = true;
+
+    final upgradeAlert = wrapper(
+      UpgradeAlert(
+        upgrader: upgrader,
+        cupertinoButtonTextStyle: cupertinoButtonTextStyle,
+        dialogStyle: UpgradeDialogStyle.cupertino,
+        onUpdate: () {
+          called = true;
+          return true;
+        },
+        onIgnore: () {
+          notCalled = false;
+          return true;
+        },
+        onLater: () {
+          notCalled = false;
+          return true;
+        },
+        child: const Center(child: Text('Upgrading')),
+      ),
+    );
+    await tester.pumpWidget(upgradeAlert);
 
     expect(find.text('Upgrader test'), findsOneWidget);
     expect(find.text('Upgrading'), findsOneWidget);
 
     // Pump the UI so the upgrader can display its dialog
     await tester.pumpAndSettle();
-    await tester.pumpAndSettle();
 
     expect(upgrader.isTooSoon(), true);
 
-    expect(find.text(upgrader.messages.title), findsOneWidget);
-    expect(find.text(upgrader.message()), findsOneWidget);
-    expect(find.text(upgrader.messages.releaseNotes), findsOneWidget);
+    expect(find.text(upgrader.state.messages!.title), findsOneWidget);
+    expect(find.text(upgrader.body(upgrader.state.messages!)), findsOneWidget);
+    expect(find.text(upgrader.state.messages!.releaseNotes), findsOneWidget);
     expect(find.text(upgrader.releaseNotes!), findsOneWidget);
-    expect(find.text(upgrader.messages.prompt), findsOneWidget);
+    expect(find.text(upgrader.state.messages!.prompt), findsOneWidget);
     expect(find.byType(CupertinoDialogAction), findsNWidgets(3));
-    expect(find.text(upgrader.messages.buttonTitleIgnore), findsOneWidget);
-    expect(find.text(upgrader.messages.buttonTitleLater), findsOneWidget);
-    expect(find.text(upgrader.messages.buttonTitleUpdate), findsOneWidget);
+    expect(
+      find.byWidgetPredicate((widget) =>
+          widget is CupertinoDialogAction &&
+          widget.textStyle == cupertinoButtonTextStyle),
+      findsNWidgets(3),
+    );
+    expect(
+        find.text(upgrader.state.messages!.buttonTitleIgnore), findsOneWidget);
+    expect(
+        find.text(upgrader.state.messages!.buttonTitleLater), findsOneWidget);
+    expect(
+        find.text(upgrader.state.messages!.buttonTitleUpdate), findsOneWidget);
+    expect(find.byKey(const Key('upgrader_alert_dialog')), findsOneWidget);
 
-    await tester.tap(find.text(upgrader.messages.buttonTitleUpdate));
+    await tester.tap(find.text(upgrader.state.messages!.buttonTitleUpdate));
     await tester.pumpAndSettle();
-    expect(find.text(upgrader.messages.buttonTitleIgnore), findsNothing);
-    expect(find.text(upgrader.messages.buttonTitleLater), findsNothing);
-    expect(find.text(upgrader.messages.buttonTitleUpdate), findsNothing);
+    expect(find.text(upgrader.state.messages!.buttonTitleIgnore), findsNothing);
+    expect(find.text(upgrader.state.messages!.buttonTitleLater), findsNothing);
+    expect(find.text(upgrader.state.messages!.buttonTitleUpdate), findsNothing);
     expect(called, true);
     expect(notCalled, true);
   }, skip: false);
-  testWidgets('test UpgradeWidget ignore', (WidgetTester tester) async {
+
+  testWidgets('test UpgradeAlert ignore', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader(platform: TargetPlatform.iOS, client: client);
+    final upgrader = Upgrader(
+      upgraderOS: MockUpgraderOS(ios: true),
+      client: client,
+      debugLogging: true,
+    );
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -241,41 +427,54 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
-
-    var called = false;
-    var notCalled = true;
-    upgrader.onIgnore = () {
-      called = true;
-      return true;
-    };
-    upgrader.onUpdate = () {
-      notCalled = false;
-      return true;
-    };
-    upgrader.onLater = () {
-      notCalled = false;
-      return true;
-    };
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     expect(upgrader.isTooSoon(), false);
 
-    await tester.pumpWidget(_MyWidget(upgrader: upgrader));
+    expect(upgrader.state.messages, isNull);
+    upgrader.updateState(upgrader.state.copyWith(messages: UpgraderMessages()));
+    expect(upgrader.state.messages, isNotNull);
+
+    var called = false;
+    var notCalled = true;
+    final upgradeAlert = wrapper(
+      UpgradeAlert(
+        upgrader: upgrader,
+        onUpdate: () {
+          notCalled = false;
+          return true;
+        },
+        onIgnore: () {
+          called = true;
+          return true;
+        },
+        onLater: () {
+          notCalled = false;
+          return true;
+        },
+        child: const Center(child: Text('Upgrading')),
+      ),
+    );
+    await tester.pumpWidget(upgradeAlert);
 
     // Pump the UI so the upgrader can display its dialog
     await tester.pumpAndSettle();
-    await tester.pumpAndSettle();
 
-    await tester.tap(find.text(upgrader.messages.buttonTitleIgnore));
+    await tester.tap(find.text(upgrader.state.messages!.buttonTitleIgnore));
     await tester.pumpAndSettle();
-    expect(find.text(upgrader.messages.buttonTitleIgnore), findsNothing);
+    expect(find.text(upgrader.state.messages!.buttonTitleIgnore), findsNothing);
     expect(called, true);
     expect(notCalled, true);
   }, skip: false);
 
-  testWidgets('test UpgradeWidget later', (WidgetTester tester) async {
+  testWidgets('test UpgradeAlert later', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader(platform: TargetPlatform.iOS, client: client);
+    final upgrader = Upgrader(
+      upgraderOS: MockUpgraderOS(ios: true),
+      client: client,
+      debugLogging: true,
+    );
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -283,41 +482,53 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
-
-    var called = false;
-    var notCalled = true;
-    upgrader.onLater = () {
-      called = true;
-      return true;
-    };
-    upgrader.onIgnore = () {
-      notCalled = false;
-      return true;
-    };
-    upgrader.onUpdate = () {
-      notCalled = false;
-      return true;
-    };
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     expect(upgrader.isTooSoon(), false);
 
-    await tester.pumpWidget(_MyWidget(upgrader: upgrader));
+    expect(upgrader.state.messages, isNull);
+    upgrader.updateState(upgrader.state.copyWith(messages: UpgraderMessages()));
+    expect(upgrader.state.messages, isNotNull);
+
+    var called = false;
+    var notCalled = true;
+    final upgradeAlert = wrapper(
+      UpgradeAlert(
+        upgrader: upgrader,
+        onUpdate: () {
+          notCalled = false;
+          return true;
+        },
+        onIgnore: () {
+          notCalled = false;
+          return true;
+        },
+        onLater: () {
+          called = true;
+          return true;
+        },
+        child: const Center(child: Text('Upgrading')),
+      ),
+    );
+    await tester.pumpWidget(upgradeAlert);
 
     // Pump the UI so the upgrader can display its dialog
     await tester.pumpAndSettle();
-    await tester.pumpAndSettle();
 
-    await tester.tap(find.text(upgrader.messages.buttonTitleLater));
+    await tester.tap(find.text(upgrader.state.messages!.buttonTitleLater));
     await tester.pumpAndSettle();
-    expect(find.text(upgrader.messages.buttonTitleLater), findsNothing);
+    expect(find.text(upgrader.state.messages!.buttonTitleLater), findsNothing);
     expect(called, true);
     expect(notCalled, true);
   }, skip: false);
 
-  testWidgets('test UpgradeWidget pop scope', (WidgetTester tester) async {
+  testWidgets('test UpgradeAlert pop scope', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader(platform: TargetPlatform.iOS, client: client);
+    final upgrader = Upgrader(
+        upgraderOS: MockUpgraderOS(ios: true),
+        client: client,
+        debugLogging: true);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -325,166 +536,58 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
-
-    var called = false;
-    upgrader.shouldPopScope = () {
-      called = true;
-      return true;
-    };
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     expect(upgrader.isTooSoon(), false);
 
-    await tester.pumpWidget(_MyWidget(upgrader: upgrader));
+    expect(upgrader.state.messages, isNull);
+    upgrader.updateState(upgrader.state.copyWith(messages: UpgraderMessages()));
+    expect(upgrader.state.messages, isNotNull);
+
+    var called = false;
+    final upgradeAlert = wrapper(
+      UpgradeAlert(
+        upgrader: upgrader,
+        shouldPopScope: () {
+          called = true;
+          return true;
+        },
+        child: const Center(child: Text('Upgrading')),
+      ),
+    );
+    await tester.pumpWidget(upgradeAlert);
 
     // Pump the UI so the upgrader can display its dialog
     await tester.pumpAndSettle();
-    await tester.pumpAndSettle();
 
-    // TODO: this test does not pop scope because there is no way to do that.
-    // await tester.pageBack();
-    // await tester.pumpAndSettle();
-    // expect(find.text(upgrader.messages.buttonTitleLater), findsNothing);
-    expect(called, false);
-  }, skip: false);
-
-  testWidgets('test UpgradeWidget Card upgrade', (WidgetTester tester) async {
-    final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader(
-        platform: TargetPlatform.iOS, client: client, debugLogging: true);
-
-    upgrader.installPackageInfo(
-        packageInfo: PackageInfo(
-            appName: 'Upgrader',
-            packageName: 'com.larryaasen.upgrader',
-            version: '0.9.9',
-            buildNumber: '400'));
-    await upgrader.initialize();
-
-    expect(upgrader.messages, isNotNull);
-
-    var called = false;
-    var notCalled = true;
-    upgrader.onUpdate = () {
-      called = true;
-      return true;
-    };
-    upgrader.onLater = () {
-      notCalled = false;
-      return true;
-    };
-    upgrader.onIgnore = () {
-      notCalled = false;
-      return true;
-    };
-
-    expect(upgrader.isTooSoon(), false);
-
-    await tester.pumpWidget(_MyWidgetCard(upgrader: upgrader));
-
-    // Pump the UI so the upgrade card is displayed
-    await tester.pumpAndSettle();
-
-    expect(find.text(upgrader.messages.releaseNotes), findsOneWidget);
-    expect(find.text(upgrader.releaseNotes!), findsOneWidget);
-    await tester.tap(find.text(upgrader.messages.buttonTitleUpdate));
-    await tester.pumpAndSettle();
-
+    final dynamic widgetsAppState = tester.state(find.byType(WidgetsApp));
+    await widgetsAppState.didPopRoute();
+    await tester.pump();
     expect(called, true);
-    expect(notCalled, true);
-    expect(find.text(upgrader.messages.buttonTitleUpdate), findsNothing);
-  }, skip: false);
+  });
 
-  testWidgets('test UpgradeWidget Card ignore', (WidgetTester tester) async {
-    final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader(
-        platform: TargetPlatform.iOS, client: client, debugLogging: true);
+  testWidgets('test UpgradeAlert no update', (WidgetTester tester) async {
+    expect(Upgrader.sharedInstance.isTooSoon(), false);
 
-    upgrader.installPackageInfo(
-        packageInfo: PackageInfo(
-            appName: 'Upgrader',
-            packageName: 'com.larryaasen.upgrader',
-            version: '0.9.9',
-            buildNumber: '400'));
-    await upgrader.initialize();
+    final upgradeAlert = wrapper(UpgradeAlert());
+    await tester.pumpWidget(upgradeAlert);
 
-    var called = false;
-    var notCalled = true;
-    upgrader.onIgnore = () {
-      called = true;
-      return true;
-    };
-    upgrader.onLater = () {
-      notCalled = false;
-      return true;
-    };
-    upgrader.onUpdate = () {
-      notCalled = false;
-      return true;
-    };
-
-    expect(upgrader.isTooSoon(), false);
-
-    await tester.pumpWidget(_MyWidgetCard(upgrader: upgrader));
-
-    // Pump the UI so the upgrade card is displayed
+    // Pump the UI
     await tester.pumpAndSettle();
 
-    await tester.tap(find.text(upgrader.messages.buttonTitleIgnore));
-    await tester.pumpAndSettle();
-
-    expect(called, true);
-    expect(notCalled, true);
-    expect(find.text(upgrader.messages.buttonTitleIgnore), findsNothing);
-  }, skip: false);
-
-  testWidgets('test UpgradeWidget Card later', (WidgetTester tester) async {
-    final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader(
-        platform: TargetPlatform.iOS, client: client, debugLogging: true);
-
-    upgrader.installPackageInfo(
-        packageInfo: PackageInfo(
-            appName: 'Upgrader',
-            packageName: 'com.larryaasen.upgrader',
-            version: '0.9.9',
-            buildNumber: '400'));
-    await upgrader.initialize();
-
-    var called = false;
-    var notCalled = true;
-    upgrader.onLater = () {
-      called = true;
-      return true;
-    };
-    upgrader.onIgnore = () {
-      notCalled = false;
-      return true;
-    };
-    upgrader.onUpdate = () {
-      notCalled = false;
-      return true;
-    };
-
-    expect(upgrader.isTooSoon(), false);
-
-    await tester.pumpWidget(_MyWidgetCard(upgrader: upgrader));
-
-    // Pump the UI so the upgrade card is displayed
-    await tester.pumpAndSettle(const Duration(milliseconds: 5000));
-
-    await tester.tap(find.text(upgrader.messages.buttonTitleLater));
-    await tester.pumpAndSettle();
-
-    expect(called, true);
-    expect(notCalled, true);
-    expect(find.text(upgrader.messages.buttonTitleLater), findsNothing);
-  }, skip: false);
+    expect(find.text('IGNORE'), findsNothing);
+    expect(find.text('LATER'), findsNothing);
+    expect(find.text('UPDATE'), findsNothing);
+    expect(find.text('Release Notes'), findsNothing);
+  });
 
   testWidgets('test upgrader minAppVersion', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
     final upgrader = Upgrader(
-        platform: TargetPlatform.iOS, client: client, debugLogging: true);
+        upgraderOS: MockUpgraderOS(ios: true),
+        client: client,
+        debugLogging: true);
     upgrader.minAppVersion = '1.0.0';
 
     upgrader.installPackageInfo(
@@ -493,7 +596,8 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     expect(upgrader.isTooSoon(), false);
     upgrader.minAppVersion = '0.5.0';
@@ -513,21 +617,33 @@ void main() {
 
     upgrader.minAppVersion = '1.0.0';
 
-    await tester.pumpWidget(_MyWidgetCard(upgrader: upgrader));
+    final upgradeCard = wrapper(
+      UpgradeCard(
+        upgrader: upgrader,
+      ),
+    );
+    await tester.pumpWidget(upgradeCard);
 
     // Pump the UI so the upgrade card is displayed
     await tester.pumpAndSettle(const Duration(milliseconds: 5000));
 
-    expect(find.text(upgrader.messages.buttonTitleIgnore), findsNothing);
-    expect(find.text(upgrader.messages.buttonTitleLater), findsNothing);
-    expect(find.text(upgrader.messages.buttonTitleUpdate), findsOneWidget);
+    expect(upgrader.state.messages, isNull);
+    upgrader.updateState(upgrader.state.copyWith(messages: UpgraderMessages()));
+    expect(upgrader.state.messages, isNotNull);
+
+    expect(find.text(upgrader.state.messages!.buttonTitleIgnore), findsNothing);
+    expect(find.text(upgrader.state.messages!.buttonTitleLater), findsNothing);
+    expect(
+        find.text(upgrader.state.messages!.buttonTitleUpdate), findsOneWidget);
   }, skip: false);
 
   testWidgets('test upgrader minAppVersion description android',
       (WidgetTester tester) async {
     final client = await MockPlayStoreSearchClient.setupMockClient();
     final upgrader = Upgrader(
-        platform: TargetPlatform.android, client: client, debugLogging: true);
+        upgraderOS: MockUpgraderOS(android: true),
+        client: client,
+        debugLogging: true);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -535,10 +651,36 @@ void main() {
             packageName: 'com.testing.test2',
             version: '2.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     expect(upgrader.belowMinAppVersion(), true);
-    expect(upgrader.minAppVersion, '4.5.6');
+    expect(upgrader.state.versionInfo?.minAppVersion.toString(), '4.5.6');
+  }, skip: false);
+
+  testWidgets('test upgrader store version android',
+      (WidgetTester tester) async {
+    final client = await MockPlayStoreSearchClient.setupMockClient(
+      verifyHeaders: {'header1': 'value1'},
+    );
+    final upgrader = Upgrader(
+      upgraderOS: MockUpgraderOS(android: true),
+      client: client,
+      clientHeaders: {'header1': 'value1'},
+      debugLogging: true,
+    );
+
+    upgrader.installPackageInfo(
+        packageInfo: PackageInfo(
+            appName: 'Upgrader',
+            packageName: 'com.testing.test7',
+            version: '2.9.9',
+            buildNumber: '400'));
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
+
+    expect(upgrader.belowMinAppVersion(), isFalse);
+    expect(upgrader.state.versionInfo?.appStoreVersion, isNull);
   }, skip: false);
 
   testWidgets('test upgrader minAppVersion description ios',
@@ -547,7 +689,9 @@ void main() {
       description: 'Use this app. [:mav: 4.5.6]',
     );
     final upgrader = Upgrader(
-        platform: TargetPlatform.iOS, client: client, debugLogging: true);
+        upgraderOS: MockUpgraderOS(ios: true),
+        client: client,
+        debugLogging: true);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -555,19 +699,21 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '2.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     expect(upgrader.belowMinAppVersion(), true);
-    expect(upgrader.minAppVersion, '4.5.6');
+    expect(upgrader.state.versionInfo?.minAppVersion.toString(), '4.5.6');
   }, skip: false);
 
   testWidgets('test UpgradeWidget unknown app', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
     final upgrader = Upgrader(
-        platform: TargetPlatform.iOS,
+        upgraderOS: MockUpgraderOS(ios: true),
         client: client,
         debugLogging: true,
-        countryCode: 'IT');
+        countryCode: 'IT',
+        languageCode: 'en');
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -575,142 +721,267 @@ void main() {
             packageName: 'com.google.MyApp',
             version: '0.1.0',
             buildNumber: '1'));
-    await upgrader.initialize();
-
-    var called = false;
-    var notCalled = true;
-    upgrader.onLater = () {
-      called = true;
-      return true;
-    };
-    upgrader.onIgnore = () {
-      notCalled = false;
-      return true;
-    };
-    upgrader.onUpdate = () {
-      notCalled = false;
-      return true;
-    };
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     expect(upgrader.isTooSoon(), false);
 
-    await tester.pumpWidget(_MyWidgetCard(upgrader: upgrader));
+    var called = false;
+    var notCalled = true;
+    final upgradeCard = wrapper(
+      UpgradeCard(
+        upgrader: upgrader,
+        onUpdate: () {
+          notCalled = false;
+          return true;
+        },
+        onIgnore: () {
+          notCalled = false;
+          return true;
+        },
+        onLater: () {
+          called = true;
+        },
+      ),
+    );
+    await tester.pumpWidget(upgradeCard);
 
     // Pump the UI so the upgrade card is displayed
     await tester.pumpAndSettle();
 
-    final laterButton = find.text(upgrader.messages.buttonTitleLater);
+    expect(upgrader.state.messages, isNull);
+    upgrader.updateState(upgrader.state.copyWith(messages: UpgraderMessages()));
+    expect(upgrader.state.messages, isNotNull);
+
+    final laterButton = find.text(upgrader.state.messages!.buttonTitleLater);
     expect(laterButton, findsNothing);
 
     expect(called, false);
     expect(notCalled, true);
   }, skip: false);
 
-  group('initialize', () {
-    test('should use fake Appcast', () async {
-      final fakeAppcast = FakeAppcast();
-      final client = MockITunesSearchClient.setupMockClient();
-      final upgrader = Upgrader(
-          platform: TargetPlatform.iOS,
-          client: client,
-          debugLogging: true,
-          appcastConfig: fakeAppcast.config,
-          appcast: fakeAppcast)
-        ..installPackageInfo(
-          packageInfo: PackageInfo(
-            appName: 'Upgrader',
-            packageName: 'com.larryaasen.upgrader',
-            version: '1.9.6',
-            buildNumber: '42',
-          ),
-        );
+  test('should use Appcast', () async {
+    final fakeAppcast = FakeAppcast();
+    final upgrader = Upgrader(
+      upgraderOS: MockUpgraderOS(os: 'ios', ios: true),
+      debugLogging: true,
+      storeController: UpgraderStoreController(
+        oniOS: () => UpgraderAppcastStore(
+          appcastURL: 'https://sparkle-project.org/test/testappcast.xml',
+          appcast: fakeAppcast,
+        ),
+      ),
+    )..installPackageInfo(
+        packageInfo: PackageInfo(
+          appName: 'Upgrader',
+          packageName: 'com.larryaasen.upgrader',
+          version: '1.9.6',
+          buildNumber: '42',
+        ),
+      );
 
-      await upgrader.initialize();
+    await upgrader.initialize();
 
-      expect(fakeAppcast.callCount, greaterThan(0));
-    }, skip: false);
+    expect(fakeAppcast.callCount, greaterThan(0));
+  }, skip: false);
 
-    test('durationUntilAlertAgain defaults to 3 days', () async {
-      final upgrader = Upgrader();
-      expect(upgrader.durationUntilAlertAgain, const Duration(days: 3));
-    }, skip: false);
+  test('should use Appcast headers', () async {
+    final upgrader = Upgrader(
+      upgraderOS: MockUpgraderOS(ios: true),
+      debugLogging: true,
+      client: MockITunesSearchClient.setupMockClient(
+          verifyHeaders: {'header1': 'value1'}),
+      clientHeaders: {'header1': 'value1'},
+      storeController: UpgraderStoreController(
+        oniOS: () => UpgraderAppcastStore(
+          appcastURL: 'https://sparkle-project.org/test/testappcast.xml',
+        ),
+      ),
+      upgraderDevice: MockUpgraderDevice(),
+    )..installPackageInfo(
+        packageInfo: PackageInfo(
+          appName: 'Upgrader',
+          packageName: 'com.larryaasen.upgrader',
+          version: '1.9.6',
+          buildNumber: '42',
+        ),
+      );
 
-    test('durationUntilAlertAgain is 0 days', () async {
-      final upgrader =
-          Upgrader(durationUntilAlertAgain: const Duration(seconds: 0));
-      expect(upgrader.durationUntilAlertAgain, const Duration(seconds: 0));
+    await upgrader.initialize();
+  }, skip: false);
 
-      UpgradeAlert(upgrader: upgrader);
-      expect(upgrader.durationUntilAlertAgain, const Duration(seconds: 0));
+  test('will use appcast critical version if exists', () async {
+    final upgraderOS = MockUpgraderOS(android: true);
+    final Client mockClient =
+        setupMockClient(filePath: 'test/testappcast_critical.xml');
 
-      UpgradeCard(upgrader: upgrader);
-      expect(upgrader.durationUntilAlertAgain, const Duration(seconds: 0));
-    }, skip: false);
+    final upgrader = Upgrader(
+      client: mockClient,
+      upgraderOS: upgraderOS,
+      upgraderDevice: MockUpgraderDevice(),
+      debugLogging: true,
+      storeController: UpgraderStoreController(
+        onAndroid: () => UpgraderAppcastStore(
+          appcastURL: 'https://sparkle-project.org/test/testappcast.xml',
+          // client: mockClient,
+        ),
+      ),
+    )..installPackageInfo(
+        packageInfo: PackageInfo(
+          appName: 'Upgrader',
+          packageName: 'com.larryaasen.upgrader',
+          version: '1.9.6',
+          buildNumber: '42',
+        ),
+      );
 
-    test('durationUntilAlertAgain card is valid', () async {
-      final upgrader =
-          Upgrader(durationUntilAlertAgain: const Duration(days: 3));
-      UpgradeCard(upgrader: upgrader);
-      expect(upgrader.durationUntilAlertAgain, const Duration(days: 3));
+    await upgrader.initialize();
 
-      final upgrader2 =
-          Upgrader(durationUntilAlertAgain: const Duration(days: 10));
-      final _ = UpgradeCard(upgrader: upgrader2);
-      expect(upgrader2.durationUntilAlertAgain, const Duration(days: 10));
-    }, skip: false);
+    var notCalled = true;
+    upgrader.willDisplayUpgrade = ({
+      required bool display,
+      String? installedVersion,
+      UpgraderVersionInfo? versionInfo,
+    }) {
+      expect(display, true);
+      expect(installedVersion, '1.9.6');
 
-    test('durationUntilAlertAgain alert is valid', () async {
-      final upgrader =
-          Upgrader(durationUntilAlertAgain: const Duration(days: 3));
-      UpgradeAlert(upgrader: upgrader);
-      expect(upgrader.durationUntilAlertAgain, const Duration(days: 3));
+      /// Appcast Test critical version.
+      expect(versionInfo!.appStoreVersion.toString(), '3.0.0');
+      notCalled = false;
+    };
 
-      final upgrader2 =
-          Upgrader(durationUntilAlertAgain: const Duration(days: 10));
-      UpgradeAlert(upgrader: upgrader2);
-      expect(upgrader2.durationUntilAlertAgain, const Duration(days: 10));
-    }, skip: false);
-  });
+    final shouldDisplayUpgrade = upgrader.shouldDisplayUpgrade();
+
+    expect(shouldDisplayUpgrade, isTrue);
+    expect(notCalled, false);
+  }, skip: false);
+
+  test('will use appcast last item', () async {
+    final upgraderOS = MockUpgraderOS(ios: true);
+
+    final Client mockClient =
+        setupMockClient(filePath: 'test/testappcastmulti.xml');
+
+    final upgrader = Upgrader(
+      client: mockClient,
+      upgraderOS: upgraderOS,
+      upgraderDevice: MockUpgraderDevice(),
+      debugLogging: true,
+      storeController: UpgraderStoreController(
+        oniOS: () => UpgraderAppcastStore(
+            appcastURL: 'https://sparkle-project.org/test/testappcast.xml'),
+      ),
+    )..installPackageInfo(
+        packageInfo: PackageInfo(
+          appName: 'Upgrader',
+          packageName: 'com.larryaasen.upgrader',
+          version: '1.9.6',
+          buildNumber: '42',
+        ),
+      );
+
+    await upgrader.initialize();
+
+    var notCalled = true;
+    upgrader.willDisplayUpgrade = ({
+      required bool display,
+      String? installedVersion,
+      UpgraderVersionInfo? versionInfo,
+    }) {
+      expect(display, true);
+      expect(installedVersion, '1.9.6');
+      expect(versionInfo!.appStoreVersion.toString(), '2.3.2');
+      notCalled = false;
+    };
+
+    final shouldDisplayUpgrade = upgrader.shouldDisplayUpgrade();
+
+    expect(shouldDisplayUpgrade, isTrue);
+    expect(notCalled, false);
+  }, skip: false);
+
+  test('durationUntilAlertAgain defaults to 3 days', () async {
+    final upgrader = Upgrader();
+    expect(upgrader.state.durationUntilAlertAgain, const Duration(days: 3));
+  }, skip: false);
+
+  test('durationUntilAlertAgain is 0 days', () async {
+    final upgrader =
+        Upgrader(durationUntilAlertAgain: const Duration(seconds: 0));
+    expect(upgrader.state.durationUntilAlertAgain, const Duration(seconds: 0));
+
+    UpgradeAlert(upgrader: upgrader);
+    expect(upgrader.state.durationUntilAlertAgain, const Duration(seconds: 0));
+
+    UpgradeCard(upgrader: upgrader);
+    expect(upgrader.state.durationUntilAlertAgain, const Duration(seconds: 0));
+  }, skip: false);
+
+  test('durationUntilAlertAgain card is valid', () async {
+    final upgrader = Upgrader(durationUntilAlertAgain: const Duration(days: 3));
+    UpgradeCard(upgrader: upgrader);
+    expect(upgrader.state.durationUntilAlertAgain, const Duration(days: 3));
+
+    final upgrader2 =
+        Upgrader(durationUntilAlertAgain: const Duration(days: 10));
+    UpgradeCard(upgrader: upgrader2);
+    expect(upgrader2.state.durationUntilAlertAgain, const Duration(days: 10));
+  }, skip: false);
+
+  test('durationUntilAlertAgain alert is valid', () async {
+    final upgrader = Upgrader(durationUntilAlertAgain: const Duration(days: 3));
+    UpgradeAlert(upgrader: upgrader);
+    expect(upgrader.state.durationUntilAlertAgain, const Duration(days: 3));
+
+    final upgrader2 =
+        Upgrader(durationUntilAlertAgain: const Duration(days: 10));
+    UpgradeAlert(upgrader: upgrader2);
+    expect(upgrader2.state.durationUntilAlertAgain, const Duration(days: 10));
+  }, skip: false);
 
   group('shouldDisplayUpgrade', () {
-    test('should respect debugDisplayAlways property', () {
+    test('should respect debugDisplayAlways property', () async {
       final client = MockITunesSearchClient.setupMockClient();
       final upgrader = Upgrader(
-          platform: TargetPlatform.iOS, client: client, debugLogging: true);
+          upgraderDevice: MockUpgraderDevice(),
+          upgraderOS: MockUpgraderOS(ios: true),
+          client: client,
+          debugLogging: true);
 
       expect(upgrader.shouldDisplayUpgrade(), false);
-      upgrader.debugDisplayAlways = true;
+      upgrader.updateState(upgrader.state.copyWith(debugDisplayAlways: true));
       expect(upgrader.shouldDisplayUpgrade(), true);
-      upgrader.debugDisplayAlways = false;
+      upgrader.updateState(upgrader.state.copyWith(debugDisplayAlways: false));
       expect(upgrader.shouldDisplayUpgrade(), false);
 
       // Test the willDisplayUpgrade callback
       var notCalled = true;
-      upgrader.willDisplayUpgrade = (
-          {required bool display,
-          String? minAppVersion,
-          String? installedVersion,
-          String? appStoreVersion}) {
+      upgrader.willDisplayUpgrade = ({
+        required bool display,
+        String? installedVersion,
+        UpgraderVersionInfo? versionInfo,
+      }) {
         expect(display, false);
-        expect(minAppVersion, isNull);
+        expect(versionInfo?.minAppVersion, isNull);
         expect(installedVersion, isNull);
-        expect(appStoreVersion, isNull);
+        expect(versionInfo?.appStoreVersion, isNull);
         notCalled = false;
       };
       expect(upgrader.shouldDisplayUpgrade(), false);
       expect(notCalled, false);
 
-      upgrader.debugDisplayAlways = true;
+      upgrader.updateState(upgrader.state.copyWith(debugDisplayAlways: true));
       notCalled = true;
-      upgrader.willDisplayUpgrade = (
-          {required bool display,
-          String? minAppVersion,
-          String? installedVersion,
-          String? appStoreVersion}) {
+      upgrader.willDisplayUpgrade = ({
+        required bool display,
+        String? installedVersion,
+        UpgraderVersionInfo? versionInfo,
+      }) {
         expect(display, true);
-        expect(minAppVersion, isNull);
+        expect(versionInfo?.minAppVersion, isNull);
         expect(installedVersion, isNull);
-        expect(appStoreVersion, isNull);
+        expect(versionInfo?.appStoreVersion, isNull);
         notCalled = false;
       };
       expect(upgrader.shouldDisplayUpgrade(), true);
@@ -719,9 +990,12 @@ void main() {
 
     test('should return true when version is below minAppVersion', () async {
       final upgrader = Upgrader(
-          debugLogging: true,
-          platform: TargetPlatform.iOS,
-          client: MockITunesSearchClient.setupMockClient())
+        debugLogging: true,
+        upgraderOS: MockUpgraderOS(ios: true),
+        client: MockITunesSearchClient.setupMockClient(
+            verifyHeaders: {'header1': 'value1'}),
+        clientHeaders: {'header1': 'value1'},
+      )
         ..minAppVersion = '2.0.0'
         ..installPackageInfo(
           packageInfo: PackageInfo(
@@ -734,16 +1008,16 @@ void main() {
 
       await upgrader.initialize();
       var notCalled = true;
-      upgrader.willDisplayUpgrade = (
-          {required bool display,
-          String? minAppVersion,
-          String? installedVersion,
-          String? appStoreVersion}) {
+      upgrader.willDisplayUpgrade = ({
+        required bool display,
+        String? installedVersion,
+        UpgraderVersionInfo? versionInfo,
+      }) {
         expect(display, true);
-        expect(minAppVersion, '2.0.0');
+        expect(versionInfo!.minAppVersion, isNull);
         expect(upgrader.minAppVersion, '2.0.0');
         expect(installedVersion, '1.9.6');
-        expect(appStoreVersion, '5.6');
+        expect(versionInfo.appStoreVersion.toString(), '5.6.0');
         notCalled = false;
       };
 
@@ -756,7 +1030,7 @@ void main() {
     test('should return true when bestItem has critical update', () async {
       final upgrader = Upgrader(
           debugLogging: true,
-          platform: TargetPlatform.iOS,
+          upgraderOS: MockUpgraderOS(ios: true),
           client: MockITunesSearchClient.setupMockClient())
         ..installPackageInfo(
           packageInfo: PackageInfo(
@@ -777,7 +1051,7 @@ void main() {
     test('packageInfo is empty', () async {
       final upgrader = Upgrader(
           client: MockITunesSearchClient.setupMockClient(),
-          platform: TargetPlatform.iOS,
+          upgraderOS: MockUpgraderOS(ios: true),
           debugLogging: true)
         ..installPackageInfo(
           packageInfo: PackageInfo(
@@ -791,14 +1065,76 @@ void main() {
       await upgrader.initialize();
       expect(upgrader.shouldDisplayUpgrade(), isFalse);
       expect(upgrader.appName(), isEmpty);
-      expect(upgrader.currentInstalledVersion(), isEmpty);
+      expect(upgrader.currentInstalledVersion, isEmpty);
     }, skip: false);
+
+    testWidgets('test UpgradeAlert with GoRouter', (WidgetTester tester) async {
+      final client = MockITunesSearchClient.setupMockClient();
+      final upgrader = Upgrader(
+        upgraderOS: MockUpgraderOS(ios: true),
+        client: client,
+        debugLogging: true,
+      );
+
+      upgrader.installPackageInfo(
+          packageInfo: PackageInfo(
+              appName: 'Upgrader',
+              packageName: 'com.larryaasen.upgrader',
+              version: '0.9.9',
+              buildNumber: '400'));
+
+      GoRouter routerConfig = GoRouter(
+        initialLocation: '/page2',
+        routes: [
+          GoRoute(
+            path: '/page1',
+            builder: (BuildContext context, GoRouterState state) => Scaffold(
+                appBar: AppBar(title: const Text('Upgrader GoRouter Example')),
+                body: const Center(child: Text('Checking... page1'))),
+          ),
+          GoRoute(
+            path: '/page2',
+            builder: (BuildContext context, GoRouterState state) => Scaffold(
+              appBar: AppBar(title: const Text('Upgrader GoRouter Example')),
+              body: const Center(child: Text('Checking... page2')),
+            ),
+          ),
+        ],
+      );
+
+      final router = MaterialApp.router(
+        title: 'Upgrader GoRouter Example',
+        routerConfig: routerConfig,
+        builder: (context, child) {
+          return UpgradeAlert(
+            upgrader: upgrader,
+            navigatorKey: routerConfig.routerDelegate.navigatorKey,
+            child: child ?? const Text('child'),
+          );
+        },
+      );
+
+      await tester.pumpWidget(router);
+
+      // Pump the UI so the upgrade card is displayed
+      await tester.pumpAndSettle();
+      await tester.pumpAndSettle();
+
+      expect(find.text('Upgrader GoRouter Example'), findsOneWidget);
+
+      expect(find.text('Update App?'), findsOneWidget);
+      expect(find.text('IGNORE'), findsOneWidget);
+      expect(find.text('LATER'), findsOneWidget);
+      expect(find.text('UPDATE NOW'), findsOneWidget);
+      expect(find.text('Release Notes'), findsOneWidget);
+    });
   });
 
   test('test UpgraderMessages', () {
     verifyMessages(UpgraderMessages(code: 'en'), 'en');
     verifyMessages(UpgraderMessages(code: 'ar'), 'ar');
     verifyMessages(UpgraderMessages(code: 'bn'), 'bn');
+    verifyMessages(UpgraderMessages(code: 'da'), 'da');
     verifyMessages(UpgraderMessages(code: 'es'), 'es');
     verifyMessages(UpgraderMessages(code: 'fa'), 'fa');
     verifyMessages(UpgraderMessages(code: 'fil'), 'fil');
@@ -815,6 +1151,7 @@ void main() {
     verifyMessages(UpgraderMessages(code: 'kk'), 'kk');
     verifyMessages(UpgraderMessages(code: 'km'), 'km');
     verifyMessages(UpgraderMessages(code: 'ko'), 'ko');
+    verifyMessages(UpgraderMessages(code: 'ku'), 'ku');
     verifyMessages(UpgraderMessages(code: 'lt'), 'lt');
     verifyMessages(UpgraderMessages(code: 'mn'), 'mn');
     verifyMessages(UpgraderMessages(code: 'nb'), 'nb');
@@ -843,55 +1180,16 @@ void verifyMessages(UpgraderMessages messages, String code) {
   expect(messages.message(UpgraderMessage.title), isNotEmpty);
 }
 
-class _MyWidget extends StatelessWidget {
-  final Upgrader upgrader;
-  const _MyWidget({Key? key, required this.upgrader}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Upgrader test',
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('Upgrader test'),
-        ),
-        body: UpgradeAlert(
-            upgrader: upgrader,
-            child: Column(
-              children: const <Widget>[Text('Upgrading')],
-            )),
-      ),
-    );
-  }
-}
-
-class _MyWidgetCard extends StatelessWidget {
-  final Upgrader upgrader;
-  const _MyWidgetCard({Key? key, required this.upgrader}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Upgrader test',
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('Upgrader test'),
-        ),
-        body: Column(
-          children: <Widget>[UpgradeCard(upgrader: upgrader)],
-        ),
-      ),
-    );
-  }
-}
-
 class MyUpgraderMessages extends UpgraderMessages {
   @override
   String get buttonTitleIgnore => 'aaa';
+
   @override
   String get buttonTitleLater => 'bbb';
+
   @override
   String get buttonTitleUpdate => 'ccc';
+
   @override
   String get releaseNotes => 'ddd';
 }
